Workaround horizon fails when osprofiler installed
There is a very long and boring story. This bug was reported 2 years ago,
moved to in progress and seems like never fixed. Due to the very interesting
way how everything is working it's not so often reproducible. In our case,
it's really failing. So, we shouldn't have osprofiler installed with Horizon
now, but new version of python-mistralclient that horizon depends on introduced
osprofiler as direct dependency and so horizon started failing. This workaround
doesn't disable or break osprofiler, it's just making very safe (and duty, yes)
change to make it working in any case.

Change-Id: If3e30e065169c1b1a1514dc67b887be294e6fb54